Refactor main and modularize
